Adjust delete link style and hide update button initially.
Updated the "Delete" link to include a confirmation class for better UX. Adjusted the layout by reducing margin space and hiding the "Update CMIds" button by default. These changes enhance usability and visual alignment.

+++ src/FileCabinet/SuiteScripts/Lib/Suitelet/rxrs_sl_add_credit_memo.js
@@ -579,8 +579,7 @@ define([
                      <td id="cmId" style='border: 1px solid #ccc; padding: 8px;display: none;'>${item.id}</td>
                       <td style='border: 1px solid #ccc; padding: 8px;'>
                         <a href='${viewCMLink}' style='margin-right: 10px; text-decoration: none; color: blue;'>Edit</a>|   
-                        <a href='${deleteCMSuiteletUrl}' style='margin-right: 10px; text-decoration: none; color: blue;'>     Delete</a> 
-                 
+     <a href="${deleteCMSuiteletUrl}" class="confirm-link" style='margin-right: 10px; text-decoration: none; color: blue;'>Delete</a>
                     </td>
                 `;
         innerHTML += `</tr>`;
@@ -608,8 +607,9 @@ define([
           </table>
              
       </div>
-      <div style="margin-top: 120px;">
+      <div style="margin-top: 50px;">
     <button id="updateCMIds" style="
+         display: none; /* Initially hidden */
         background-color: #007bff;
         color: white;
         border: none;
@@ -620,7 +620,7 @@ define([
         border-radius: 5px;
         transition: background-color 0.3s ease, transform 0.1s ease;
         box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.2);
-        display: inline-block;
+        
         text-align: center;"
         onmouseover="this.style.backgroundColor='#0056b3';"
         onmouseout="this.style.backgroundColor='#007bff';"
